Fix single user mode, #154961
svn path=/trunk/; revision=2382

+++ sbin/runscript.sh
@@ -669,8 +669,11 @@ for arg in $* ; do
 				fi
 			done
 			rm -rf "${svcdir}/snapshot/$$"
-		else
+		elif service_stopped "${SVCNAME}" ; then
 			rm -f "${svcdir}"/scheduled/*/"${SVCNAME}"
+			if [[ ${SOFTLEVEL} != "single" ]] ; then
+				rm -f "${svcdir}/coldplugged/${SVCNAME}"
+			fi
 		fi
 
 		;;
